refactor: extract vaadin-tab and vaadin-tabs mixins
