+++ packages/tabs/package.json
@@ -35,6 +35,7 @@
     "polymer"
   ],
   "dependencies": {
+    "@open-wc/dedupe-mixin": "^1.3.0",
     "@polymer/polymer": "^3.0.0",
     "@vaadin/a11y-base": "24.3.0-alpha10",
     "@vaadin/component-base": "24.3.0-alpha10",
